requirements.txt
numpy==1.19.2
streamlit==0.77.0
tensorflow==2.5.0
tensorflow_hub==0.9.0
matplotlib==3.0.3
Pillow==8.3.1
